move generalized model to overpass fn to sutils
